consistency changes
